change role about store, update and delete users

+++ app/Http/Controllers/LoginController.php
@@ -213,33 +213,42 @@ class LoginController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = User::where('id', '=', $id)->first();
+        if (Auth::user()->role_id == 1) {
 
-        $user->firstName = $request->firstName;
-        $user->lastName = $request->lastName;
-        $user->email = $request->email;
-        $user->password = bcrypt($request->password);
-        $user->role_id = $request->role_id;
-        $user->company_id = $request->company_id;
+            $user = User::where('id', '=', $id)->first();
 
-        if ($request->hasFile('userPicture')) {
-            Storage::disk('public')->put('userImages', $request->file('userPicture'));
-            $name = Storage::disk('public')->put('userImages', $request->file('userPicture'));
-            $user->userPicturePath = $name;
-            $user->userPictureName = $request->file('userPicture')->getClientOriginalName();
-        }
+            $user->firstName = $request->firstName;
+            $user->lastName = $request->lastName;
+            $user->email = $request->email;
+            $user->password = bcrypt($request->password);
+            $user->role_id = $request->role_id;
+            $user->company_id = $request->company_id;
 
-        if ($user->save()) {
-            return response()->json([
-                'success' => true,
-                'status' => 200,
-                'data' => $user,
-            ]);
+            if ($request->hasFile('userPicture')) {
+                Storage::disk('public')->put('userImages', $request->file('userPicture'));
+                $name = Storage::disk('public')->put('userImages', $request->file('userPicture'));
+                $user->userPicturePath = $name;
+                $user->userPictureName = $request->file('userPicture')->getClientOriginalName();
+            }
+
+            if ($user->save()) {
+                return response()->json([
+                    'success' => true,
+                    'status' => 200,
+                    'data' => $user,
+                ]);
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'status' => 500,
+                    'data' => []
+                ]);
+            }
         } else {
             return response()->json([
                 'success' => false,
-                'status' => 500,
-                'data' => []
+                'status' => 401,
+                'data' => ''
             ]);
         }
     }
@@ -252,13 +261,23 @@ class LoginController extends Controller
      */
     public function destroy($id)
     {
+
         $userDelete = User::findOrFail($id);
 
-        if ($userDelete->delete()) {
+        if (Auth::user()->role_id == 1) {
+
+            if ($userDelete->delete()) {
+                return response()->json([
+                    'success' => true,
+                    'status' => 200,
+                    'message' => 'Proof! Your User has been deleted!',
+                ]);
+            }
+        } else {
             return response()->json([
-                'success' => true,
-                'status' => 200,
-                'message' => 'Proof! Your User has been deleted!',
+                'success' => false,
+                'status' => 401,
+                'data' => ''
             ]);
         }
     }
